+++ src/const/base.ts
@@ -1,0 +1,3 @@
+export enum Symb {
+  ALL = "ALL",
+}
