+++ src/types/global.d.ts
@@ -1,0 +1,5 @@
+interface Window {
+  remote: any;
+  clipboard: any;
+  ipcRenderer: any;
+}
